requirements.txt
Faker~=37.1.0
Django~=5.2
djangorestframework~=3.16.0
factory_boy~=3.3.1
PyJWT~=2.10.1
gunicorn
psycopg2-binary
django-cors-headers
drf-yasg